+++ src/components/SelectorTap.js
